# DATN_BE/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quản Lý Mầm Non API",
        "description": "Tài liệu API cho hệ thống Quản Lý Mầm Non",
        "contact": {
            "email": "support@qlmn.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/hoc-sinh/data": {
            "get": {
                "tags": [
                    "Admin - Học sinh"
                ],
                "summary": "Lấy danh sách học sinh",
                "operationId": "a7ae41017f2e03f30592237cd10cddaa",
                "responses": {
                    "200": {
                        "description": "Lấy danh sách học sinh thành công (status = true)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Load dữ liệu học sinh thành công"
                                        },
                                        "id_chuc_nang": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "ho_va_ten": {
                                                        "type": "string",
                                                        "example": "Nguyễn Văn C"
                                                    },
                                                    "gioi_tinh": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "avatar.jpg",
                                                        "nullable": true
                                                    },
                                                    "ngay_sinh": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2018-01-01"
                                                    },
                                                    "dia_chi": {
                                                        "type": "string",
                                                        "example": "123 Đường ABC"
                                                    },
                                                    "tinh_trang": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_lop_hoc": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_phu_huynh": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_block": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "ten_lop": {
                                                        "type": "string",
                                                        "example": "Lớp Mầm 1"
                                                    },
                                                    "ten_phu_huynh": {
                                                        "type": "string",
                                                        "example": "Nguyễn Văn D"
                                                    },
                                                    "sdt_phu_huynh": {
                                                        "type": "string",
                                                        "example": "0901234567"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}